# authz: cache prometheus metric lookup in subrepo